# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "My project documents",
    "description": "",
    "contact": {
      "email": "test@mail.ua",
      "name": "Dima"
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "userLogin",
        "description": "Login user",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "default": "example@mail.com",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "default": "HelloWorld1234_!",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User was login",
            "schema": {
              "$ref": "#/definitions/loginUser"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "loginUser": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "12345some",
          "description": "MongoDB ID"
        },
        "email": {
          "type": "string",
          "example": "example@email.com"
        },
        "name": {
          "type": "string",
          "example": "Tom"
        },
        "role": {
          "type": "string",
          "example": "user"
        },
        "createdAt": {
          "type": "string",
          "example": "XXXX-XX-XXTXX:XX:XX.XXXZ"
        },
        "updatedAt": {
          "type": "string",
          "example": "XXXX-XX-XXTXX:XX:XX.XXXZ"
        },
        "__v": {
          "type": "number",
          "example": 0
        },
        "age": {
          "type": "number",
          "example": 18
        },
        "access_token": {
          "type": "string",
          "example": "djvevoeeui58392fh35u63-05unefvfj38"
        },
        "refresh_token": {
          "type": "string",
          "example": "vmdfjpoj57epdj209jf;d;lsvsdvob9rii8"
        }
      }
    },
    "commonError": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "something wrong"
        },
        "code": {
          "type": "number",
          "example": 400
        }
      }
    }
  }
}